requirements.txt
# ─── Core ───
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
ffmpeg-python>=0.2.0

# ─── UI ───
streamlit>=1.30.0

# ─── Audio Analysis ───
librosa>=0.10.0
soundfile>=0.12.0

# ─── TTS ───
edge-tts>=6.1.0

# ─── AI / Emotion (optional — rule backend works without these) ───
# openai>=1.0.0
# openai-whisper>=20230918

# ─── Dev ───
# pytest>=7.0
# black>=23.0
# ruff>=0.1.0
